baitap-submit/dan-dinh/07-function-calling-advanced/requirements.txt
groq==0.15.0
python-dotenv==0.21.0
gradio==5.12.0
pydantic==2.10.5
httpx==0.27.2